+++ Modules/core/View/Coretiles/selfjoinspaceAction.php
@@ -1,0 +1,13 @@
+<?php include 'Modules/core/View/layout.php' ?>
+
+<!-- body -->     
+<?php startblock('content') ?>
+
+<div class=" container">
+    <div class="">
+    <?php echo $form ?> 
+    </div>
+
+</div>
+<?php
+endblock();